Code refactor
Move modal code to ModalView.js
