+++ src/subpages/cryptocurrencySubpage/CryptoCurrency.js
@@ -1,17 +1,18 @@
 import provider from "../../services/Provider"
 import css from "./CryptoCurrency.css"
 import Subpage from "../Subpage"
+import ModalView from "./ModalView"
 
 export default class CryptoCurrencySubpage extends Subpage {
   _allCoins = []
   _coinsToRender = []
   _actuallyDisplayedCoinsId = []
-  _coinId
   pageSize = 80
 
   constructor() {
     super(css)
     this._currencyAPI = provider.get("CurrencyAPI")
+    this.modal = new ModalView()
   }
 
   coinsToRender = () => {
@@ -51,40 +52,6 @@ export default class CryptoCurrencySubpage extends Subpage {
     })
   }
 
-  openModal = async () => {
-    const closeModal = document.querySelector(".modal__close")
-
-    await this._actuallyDisplayedCoinsId.forEach((coin) =>
-      coin.addEventListener("click", () => {
-        this._coinId = coin.getAttribute("data-coinId")
-        this._currencyAPI
-          .getSingleCoin(this._coinId)
-          .then((r) => this.renderModal(r))
-        this.modalWindow.classList.add("modal--open")
-      })
-    )
-    closeModal.addEventListener("click", () => {
-      this.modalWindow.classList.remove("modal--open")
-    })
-  }
-
-  exchangeCoins = (e) => {
-    e.preventDefault()
-    const moneySpan = document.querySelector("span.modal__money")
-    const coinsInputValue = document.querySelector("form.modal__form").coins
-      .value
-    this._currencyAPI
-      .exchangeCoinToUSD(this._coinId, coinsInputValue)
-      .then((r) => parseFloat(r).toFixed(2))
-      .then((r) => (moneySpan.textContent = r))
-  }
-
-  exchangeCoinToUSD = async (coinId, amountOfCoins) => {
-    await this._currencyAPI
-      .exchangeCoinToUSD(coinId, amountOfCoins)
-      .then((r) => r)
-  }
-
   renderPage = () => {
     this.coinsList.innerHTML = null
     this._coinsToRender.forEach((coin) => {
@@ -96,27 +63,17 @@ ${coin.name}</p>`
     this.updateActuallyCoinId()
   }
 
-  renderModal = (clickedCoin) => {
-    const coinTitle = document.querySelector("h3.modal__title")
-    const coinSymbol = document.querySelector("span.modal__symbol")
-    const coinDescription = document.querySelector("p.modal__description")
-
-    coinTitle.textContent = clickedCoin.name
-    coinSymbol.textContent = clickedCoin.symbol
-    coinDescription.textContent = clickedCoin.description
-  }
-
   updateActuallyCoinId = () => {
     this._actuallyDisplayedCoinsId = [
       ...document.querySelectorAll("[data-coinId]"),
     ]
-    this.openModal()
+    this.modal.openModal(this._actuallyDisplayedCoinsId, this.modalWindow)
   }
 
   catchHTMLElements = () => {
     this.modalForm = document
       .querySelector("form.modal__form")
-      .addEventListener("submit", (e) => this.exchangeCoins(e))
+      .addEventListener("submit", (e) => this.modal.exchangeCoins(e))
     this.coinsList = document.querySelector("#coins-list")
     this.modalWindow = document.querySelector(".modal")
   }
